% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbachAlpha.R
\name{cronbachAlpha}
\alias{cronbachAlpha}
\title{CTT Cronbach's Alpha:}
\usage{
cronbachAlpha(data)
}
\arguments{
\item{data}{An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions.}
}
\value{
Cronbach's Alpha value for the given graded data.
}
\description{
Calculate Cronbach's Alpha for an assessment from sample data..
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPQLdata()
temp.piql.data <- PIQLdata$courses
data.num <- temp.data$data.num

# Cronbach's Alpha
cronbachAlpha(data.num)
}
