% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttCronbachAlpha.R
\name{cttCronbachAlpha}
\alias{cttCronbachAlpha}
\title{CTT Cronbach's Alpha:}
\usage{
cttCronbachAlpha(data, booted = FALSE, nRuns = 100)
}
\arguments{
\item{data}{An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions.}

\item{booted}{Logical (default = FALSE). FALSE means no bootstrapping will be
performed. TRUE turns on the bootstrapping feature.}

\item{nRuns}{Number of random samples to use in the bootstrapping (default = 100).
For publications it is recommended that 10,000 runs be performed since sample
error goes as 1/sqrt(nRuns).}
}
\value{
When booted = FALSE (the default setting) then the straight calculated
value will be returned. If booted = TRUE, the function will output the mean and
standard deviation Cronbach's Alpha values calculated from nRuns randomly sampled
with replacement data sets from the given sample.
}
\description{
Calculate Cronbach's Alpha for an assessment from sample data..
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
temp.piql.data <- PIQLdata$courses
data.num <- temp.data$data.num

# Straight Cronbach's Alpha
cttCronbachAlpha(data.num)

# Booted using 100 random samples (default)
cttCronbachAlpha(data.num, booted = TRUE)

# Booted using 1000 random samples by manually setting nRuns
cttCronbachAlpha(data.num, booted = TRUE, nRuns = 1000)
}
