% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttDifficulty-Walkthrough.R
\name{cttDifficulty.Walkthrough}
\alias{cttDifficulty.Walkthrough}
\title{CTT Difficulty WALKTHROUGH:}
\usage{
cttDifficulty.Walkthrough(data)
}
\arguments{
\item{data}{Can be either

1) An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions,

2) The output of piql.data.select with one or multiple courses selected.}
}
\value{
The CTT Difficulty analysis as specified in the walkthough
}
\description{
Walks the user through the set up and use of calculating
Classical Test Theory item difficulty for the items on an assessment
using a given sample.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
temp.data <- piql.data.select(PIQLdata, course = 2, numBlanks.allowed = 0)

# Begin walkthrough
cttDifficulty.Walkthrough(temp.data)

}
