% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttGeneralAnalysis.R
\name{cttGeneralAnalysis}
\alias{cttGeneralAnalysis}
\title{CTT General Analysis}
\usage{
cttGeneralAnalysis(
  data,
  perc = 0.27,
  as.Percentile = FALSE,
  booted = FALSE,
  nRuns = 100,
  plotBarChart = FALSE
)
}
\arguments{
\item{data}{Can be either

1) An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions,

2) The output of piql.data.select with one or multiple courses selected.}

\item{perc}{(Default = 0.27) percentage or percentile to be used in the analysis.}

\item{as.Percentile}{(Default = FALSE) If true, then students are split on percentile
not as percentage in total score.}

\item{booted}{Logical (default = FALSE). FALSE means no bootstrapping will be
performed. TRUE turns on the bootstrapping feature.}

\item{nRuns}{Number of random samples to use in the bootstrapping (default = 100).
For publications it is recommended that 10,000 runs be performed since sample
error goes as 1/sqrt(nRuns).}

\item{plotBarChart}{(Default = FALSE). Generate barchart. If bootstrapping was
performed, then error bars wll be automatically added (+- 1 standard deviation).}
}
\value{
When booted = FALSE (the default setting) then the straight calculated
value will be returned. If booted = TRUE, the function will output the mean and
standard deviation for the CTT item difficulties calculated from nRuns randomly sampled
with replacement data sets from the given sample.
}
\description{
Performs the following analysis: Cronbach's Alpha, Item point-biserial,
Item difficulty, and Item discrimination.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
pulledData <- pulldata()

# Input can be the score matrix
cttGeneralAnalysis(data.num)
# or the output from piql.data.select()
cttGeneralAnalysis(temp.data)
# Supports Bootstrapping of all statistics
cttGeneralAnalysis(data.num, booted = TRUE)
# And will plot the results (with error bars if bootstrapped)
cttGeneralAnalysis(data.num, plotBarChart = TRUE)
cttGeneralAnalysis(data.num, booted = TRUE, plotBarChart = TRUE)

}
