% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttGeneralAnalysis-Walkthrough.R
\name{cttGeneralAnalysis.Walkthrough}
\alias{cttGeneralAnalysis.Walkthrough}
\title{CTT General Analysis WALKTHROUGH:}
\usage{
cttGeneralAnalysis.Walkthrough(data)
}
\arguments{
\item{data}{Can be either

1) An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions,

2) The output of piql.data.select with one or multiple courses selected.}
}
\value{
The Cronbach's Alpha, Item Difficulty, Item Discrimination,
 and Item Point-biserial statistics plus plots as specified in the walkthough.
}
\description{
Walks the user through the set up and use a general CTT analysis
of the assessment using a given sample.
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0, MCMR.items = NA)

# Begin walkthrough
cttGeneralAnalysis.Walkthrough(temp.data)

}
