% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttPointBiserialMultiple.R
\name{cttpointBiserial.multi}
\alias{cttpointBiserial.multi}
\title{CTT Point-biserial:}
\usage{
cttpointBiserial.multi(data, booted = FALSE, nRuns = 100, plotBarChart = FALSE)
}
\arguments{
\item{data}{An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions.}

\item{booted}{Logical (default = FALSE). FALSE means no bootstrapping will be
performed. TRUE turns on the bootstrapping feature.}

\item{nRuns}{Number of random samples to use in the bootstrapping (default = 100).
For publications it is recommended that 10,000 runs be performed since sample
error goes as 1/sqrt(nRuns).}

\item{plotBarChart}{(Default = FALSE). Generate barchart. If bootstrapping was
performed, then error bars wll be automatically added (+- 1 standard deviation).}
}
\value{
Point-biserial correlation between the total score (with focus item removed)
and the focus item. For all items on the assessment
}
\description{
Calculates the point-biserial correlation between the total score
(with focus item removed) and the focus item. This gives a measure if the items
on the assessment are all measuring a single similar latent trait. If the point-biserial
is poor, then this could be an indication that the assessment is multidimensional.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
temp.piql.data <- PIQLdata$courses
data.num <- temp.data$data.num

# Straight item difficulty.
cttpointBiserial.multi(data.num)

# Booted using 100 random samples (default)
cttpointBiserial.multi(data.num, booted = TRUE)

# Booted using 1000 random samples by manually setting nRuns
cttpointBiserial.multi(data.num, booted = TRUE, nRuns = 1000)

}
