% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectData-Walkthrough.R
\name{data.select.Walkthrough}
\alias{data.select.Walkthrough}
\title{Select specific course from the pulled data - Walkthrough.}
\usage{
data.select.Walkthrough(pulled.data)
}
\arguments{
\item{pulled.data}{Output from pullPQLdata or pullData or import.from.computer}
}
\value{
A list of course data:

     $data.alpha ALPHABETICAL data for the selected course.

     $data.num NUMERICAL data for the selected course.

     $nS.details Details about how many students were removed due to the specified numBlanks.allowed value.
}
\description{
Guides the user through selecting the correct data from the
pulled data from either AWS or from their own computer.
}
\examples{
# Select data via walkthrough using data pulled
# from AWS or personal computer.
data.select.Walkthrough(pulled.data)
}
