% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itaDiff.R
\name{diff.ita}
\alias{diff.ita}
\title{Item Tree Analysis: Error in reconstructed condtradiction matrix}
\usage{
\method{diff}{ita}(contradiction.mat.orig, contradiction.mat.cur)
}
\arguments{
\item{contradiction.mat.orig}{Output from the contradiction.mat.generator() function.
Original contradiction matrix for the data.}

\item{contradiction.mat.cur}{Current estimated contradication matrix. Calculated
differently depending on ITA method: Original, Corrected, or Monimized Corrected.}
}
\value{
The error (similar to a Chi^2) in the estimated contradiction matrix
for the current tree.
}
\description{
Calculates error between the reconstructed and original
contradiction matrices. Goal in ITA is to get this to be as small as possible.
}
