% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efaWalkthrough.R
\name{efa.walkthrough}
\alias{efa.walkthrough}
\title{Walkthrough of an Exploratory Factor Analysis}
\usage{
efa.walkthrough(data)
}
\arguments{
\item{data}{Can be either

1) An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions,

2) The output of piql.data.select with one or multiple courses selected.}
}
\value{
Walks user through an EFA of the data.
}
\description{
Uses prompts to walks the user through an EFA of the given data.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
temp.piql.data <- PIQLdata$courses
data.num <- temp.data$data.num

Begin walkthrough
efa.walkthrough(data.num)
}
