% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalSaveLocation.R
\name{ensure_library}
\alias{ensure_library}
\title{Choose Save Location}
\usage{
ensure_library(lib.name)
}
\value{
Enables users to select directories for saving analysis results.
}
\description{
Opens a directory GUI to help the user select a save location.
}
