% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itaErrorRateMinCor.R
\name{error.rate.minCor}
\alias{error.rate.minCor}
\title{Item Tree Analysis: Error Rate for minimized corrected ITA}
\usage{
error.rate.minCor(current.tree, condradiction.mat, p, nS)
}
\arguments{
\item{current.tree}{Output from the itaForest() function. Current tree being considered.}

\item{condradiction.mat}{Output from the contradiction.mat.generator() function.
Contradiction matrix for the data.}

\item{p}{CTT Item Difficulty.}

\item{nS}{Number of students in the data.}
}
\value{
Error rate for the current tree using the minimized corrected method.
}
\description{
Calculates the minimized corrected error rate between the links carried in the current
tree and the contradiction matrix.
}
