% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itaErrorRateOriginal.R
\name{error.rate.orig}
\alias{error.rate.orig}
\title{Item Tree Analysis: Error Rate for original and corrected ITA}
\usage{
error.rate.orig(current.tree, condradiction.mat, p, nS)
}
\arguments{
\item{current.tree}{Output from the itaForest() function. Current tree being considered.}

\item{condradiction.mat}{Output from the contradiction.mat.generator() function.
Contradiction matrix for the data.}

\item{p}{CTT Item Difficulty.}

\item{nS}{Number of students in the data.}
}
\value{
Error rate for the current tree.
}
\description{
Calculates the error rate between the links carried in the current
tree and the contradiction matrix.
}
