% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalExportasExcel.R
\name{export.as.excel}
\alias{export.as.excel}
\title{Export to Spreadsheet}
\usage{
export.as.excel(to.be.exported)
}
\arguments{
\item{to.be.exported}{Output from one of simplyPsycho's analysis functions.}
}
\value{
Saves an .xlsx to the current working directory.
}
\description{
Export the outputs of simplyPsycho function as an Excel Spreadsheet
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0)
results <- cttDifficulty(temp.data)

export.as.excel(results,"filename.xlsx")
}
