% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalCompareCourse.R
\name{gen.compare.courses}
\alias{gen.compare.courses}
\title{Compare courses}
\usage{
gen.compare.courses(data, makeBoxPlots = FALSE, ...)
}
\arguments{
\item{data}{MUST BE the output of piql.data.select for multiple courses selected.}

\item{makeBoxPlots}{(Default = FALSE) If TRUE, then bar plots will be made for
all given courses in a single plot.}
}
\value{
Gives the Cohen's d, an interpretation of the Cohen's d, the p-value,
the adjust p-value for multiple comparisions, and an optional barplot.
}
\description{
Use a t-test to compare total scores between two courses, and
ANOVA to compate three or more courses.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
data <- piql.data.select(PIQLdata, course = c(1,3), Matched = TRUE)

# Compare total scores...
gen.compare.courses(data, method = "bonferroni")
# ... supports makinf boxplots as well
gen.compare.courses(data, method = "bonferroni", makeBoxPlots = TRUE)
}
