% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalCompareCourse-Walkthrough.R
\name{gen.compare.courses.Walkthrough}
\alias{gen.compare.courses.Walkthrough}
\title{General Comparision of Courses WALKTHROUGH:}
\usage{
gen.compare.courses.Walkthrough(data)
}
\arguments{
\item{data}{Can be either

1) An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions,

2) The output of piql.data.select with one or multiple courses selected.}
}
\value{
Cohen's d with interpretation, p-value and adjusted p-value of difference in means,
course normalized gain, difference in item difficulty with error. On request
box plots, change in difficulty plots, Sankey plots, and gains versus pre score plots
cal all be created. Sankey plots and gains versus pre score plots require matched data.

NOTE: Use the Left and Right arrow in the Plots window to switch between the plots.
}
\description{
Walks the user through a general comparision of courses in the
 given sample.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
temp.data <- piql.data.select(PIQLdata, course = 2, numBlanks.allowed = 0)

# Begin walkthrough
gen.compare.courses.Walkthrough(temp.data)
}
