% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalSankeyPlot.R
\name{general.compare.courses.MakeSankey}
\alias{general.compare.courses.MakeSankey}
\title{Makes Sankey plot}
\usage{
general.compare.courses.MakeSankey(
  set1,
  set2,
  nQ,
  nBins = 7,
  couse.names,
  ii,
  jj
)
}
\description{
Makes Sankey plot for gen.compare.course
}
