% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalImportfromComputer.R
\name{import.from.computer}
\alias{import.from.computer}
\title{Import data from computer}
\usage{
import.from.computer()
}
\value{
Data to be used in simplyPsycho with correct formatting.
}
\description{
Select files from your computer to be used in siplyPsycho. The
function asks for descriptive information about the files (student data or
answer key). If student data, then it asks for the course it is from and for
the term. formatting is checked via build.psycho.datafile. See that function's
documentation for the formatting extected for student data.
}
\examples{
# Pull in from your computer
## Data must be formatted correctly - see documentation.

import.from.computer()

# Follow the prompts.
# Data should be formatted in the same manner as simplySampleData01 available
# in the simplyPsycho Github repository (https://github.com/PhilipEaton/simplyPsycho).
Read in file from computer and pull the last letters to check file type.
}
