% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalPointBiserial.R
\name{internal.pointBiserial}
\alias{internal.pointBiserial}
\title{CTT Point-biserial:}
\usage{
internal.pointBiserial(data)
}
\arguments{
\item{data}{An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions.}
}
\value{
Point-biserial correlation between the total score (with focus item removed)
and the focus item. For all items on the assessment
}
\description{
Calculates the point-biserial correlation between the total score
(with focus item removed) and the focus item. This gives a measure if the items
on the assessment are all measuring a single similar latent trait. If the point-biserial
is poor, then this could be an indication that the assessment is multidimensional.
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0)
data.num <- temp.data$data.num

# Point-biserial
pointBiserial(data.num)
}
