% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircget.R
\name{irc.get}
\alias{irc.get}
\title{Item Response Curve data generator:}
\usage{
irc.get(data.Alpha, data.Num, nO = NULL)
}
\arguments{
\item{data.Alpha}{An nS by nQ matrix or data frame of ALPHABETICAL data,
where nS is the number of students in the sample and nQ is the number of questions.}

\item{data.Num}{An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions.}

\item{nO}{The maximum number of options available on the assessment. Default it set to
find this value automatically given the response options present in the sample.
It is suggested you enter than value manually if the max number of options is
known. For example,  if you know item 7 has 6 options, the most out of all the
items on the assessment, then set nO = 7.}
}
\value{
Plotting data for item response curves. To be put into the ircPlot() function.
}
\description{
Calculates the plotting data needed to create item response curves
for the items on an assessment using a given sample.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPQLdata()
temp.piql.data <- PIQLdata$courses
data.alpha <- temp.data$data.alpha
data.num <- temp.data$data.num

# get irc data and plot item 2
irc.data <- irc.get(data.alpha, data.num)
irc.plot(irc.data, 2)
}
