% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircPlotErrorBars.R
\name{irc.plot.withErrorBars}
\alias{irc.plot.withErrorBars}
\title{Plot Item Response Curves with error bars:}
\usage{
irc.plot.withErrorBars(irc.data.from.booted, qq)
}
\arguments{
\item{irc.data.from.booted}{output from irc.get.Booted() function.}

\item{qq}{Question number for the generated plot.}
}
\value{
IRC plot with error bars for item number qq.
}
\description{
Given the output from irc.get.Booted() function, create an IRC plot
with error bars for a specified item.
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0)
data.alpha <- temp.data$data.alpha
data.num <- temp.data$data.num

# Get irc data with default values.
irc.data.booted <- irc.get.Booted(data.alpha,data.num)
# Plot item 2.
irc.plot.withErrorBars(irc.data.booted, 2)
# Plot item 4.
irc.plot.withErrorBounds(irc.data.booted, 4)
}
