% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt2PL.R
\name{irt2PL}
\alias{irt2PL}
\title{Item Response theory 2-parameter logistic model with bootstrapping for error estimation:}
\usage{
irt2PL(data.num, Method = 1, nRuns = 5)
}
\arguments{
\item{data.num}{An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions.}

\item{Method}{(Default = 1) Method = 1 does traditional bootstrapping for
estimations of population-level error. Method = 2 does initial parameter randomization
for estimations of sample-level error.}

\item{nRuns}{Number of random samples to use in the bootstrapping (default = 5).
For publications it is recommended that 10,000 runs be performed since sample
error goes as 1/sqrt(nRuns).}
}
\value{
Model fit and item parameter estimation mean and standard deviations for the sample.

     $fit.means Model fit mean values from bootstrapping.

     $fir.stDevs Model fit standard deviation values from bootstrapping.

     $item.means Item parameter mean values from bootstrapping.

     $item.stDevs Item parameter standard deviation values from bootstrapping.
}
\description{
Calculate Item Response Theory (IRT) 2-parameter logistic model (2PL)
for given sample. Estimation of the error in item parameters can be found using
2 methods.

Method 1: Bootstrap the data. Estimate errors in item parameters by bootstrapping sample.
This given population level error estimations.

Method 2: Randomly estimate the initial parameters in mirt. Estimate errors in
item parameters by wiggling the initial parameters. This given sample level error estimations.

WARNING: IRT is in general a very computationally expensive process,
meaning it is slow. It is recommended you begin with 5 - 10 runs to get a feel for
what the data is doing, and then upping to a larger number of runs once you are
sure the output is giving you what you want.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
temp.piql.data <- PIQLdata$courses
data.alpha <- temp.data$data.alpha
data.num <- temp.data$data.num

# Get irc data with default values.
irt2PL.Method1.res <- irt2PL(data.num)
irt2PL.Method1.res$fit.means
irt2PL.Method1.res$fit.stDevs
irt2PL.Method1.res$item.means
irt2PL.Method1.res$item.stDevs

# Cange Method and number of runs
irt2PL.Method2.res <- irt2PL(data.num, Method = 2, nRuns = 10)
irt2PL.Method2.res$fit.means
irt2PL.Method2.res$fit.stDevs
irt2PL.Method2.res$item.means
irt2PL.Method2.res$item.stDevs

}
