% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itaCycleKiller.R
\name{ita.cycle.killer}
\alias{ita.cycle.killer}
\title{Item Tree Analysis - Cycle Killer}
\usage{
ita.cycle.killer()
}
\value{
A plot after killing the detected cycle... hopefully
}
\description{
TBA
}
\examples{
# TBA
}
