% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itaFullControl.R
\name{ita.fullCon}
\alias{ita.fullCon}
\title{Item Tree Analysis - Full Control}
\usage{
ita.fullCon(data, perc = 0.2, method = 1)
}
\arguments{
\item{data}{An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions.}

\item{perc}{(Default = 0.20) Percentage of either the number of students or the
max number of contradictions to be retained in the model.}

\item{method}{(Defait = 1) Method 1 is based on the number of students and will be
the same for every bootstrapped model. Method 2 is based on the maximum number
on contradictions in the initial model and can vary a little between random samples.}
}
\value{
ITA map to be plotted using itaPlot() function.

     $model Model found from analysis.

     $diff Diff for the model.
}
\description{
Performs an ITA where the user gets to control the max amount of
contractions between a causal pairing to be retained in the model.

This file does item tree analysis with bootstrapping for error estimation:
 Relaxed the code to so that models will always
 1) includes all of the items on the instrument, and
 2) select the model with the best diff from those models.

Not really recommend for publications boasting quantitative
robustness. This method is EXTREMELY adhoc and should only be used for
qualitative analysis purposes
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0, MCMR.grading = "Selected")
data.num.MCMR <- temp.data$data.num

# Fully Controlled ITA
ita.data.fc <- ita.relaxed(data.num.MCMR)
ita.data.fc$model
ita.data.fc$diff

## PLOTTIING
# Set labels
labs <- c("Plants", "Miner", "Fish", "Hooke's Law", "Ferris",
"Inv. g", "JogAB", "SphBottle", "mkp", "Slide", "Odometer",
"Squareness", "Olive Oil", "Ch.Sph.", "Int. E", "Q&N", "Bhutan-A",
"Bhutan-C", "Bhutan-D", "Work-D", "Work-G", "EField", "Delta v")

# and plot
irt.Plot(ita.data.fc$model, labs = labs)
}
