% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itaPlot.R
\name{ita.Plot}
\alias{ita.Plot}
\title{Item Tree Analysis - Create a Plot}
\usage{
ita.Plot(current.tree, labs = NULL)
}
\arguments{
\item{current.tree}{The tree to be plotted}

\item{labs}{Labels for the tree, if any.}
}
\value{
Hesse Plot of the current tree.
}
\description{
Creates a plot given a ITA tree
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0, MCMR.items = NA)
data.alpha.MCMR <- temp.data$data.alpha
data.num.MCMR <- temp.data$data.num
# Check number of student removed. Should be less than 10\%.
temp.data$nS.details

# Original ITA
ita.data.1 <- ita.strict(data.num.MCMR,type = 1)
ita.data.1$model
ita.data.1$diff
ita.data.1$errorRate

# Corrected ITA
ita.data.2 <- ita.strict(data.num.MCMR,type = 2)
ita.data.2$model
ita.data.2$diff
ita.data.2$errorRate

# Minimized Corredted ITA
ita.data.3 <- ita.strict(data.num.MCMR,type = 3)
ita.data.3$model
ita.data.3$diff
ita.data.3$errorRate

## PLOTTIING
# Set labels
labs <- c("Plants", "Miner", "Fish", "Hooke's Law", "Ferris",
"Inv. g", "JogAB", "SphBottle", "mkp", "Slide", "Odometer",
"Squareness", "Olive Oil", "Ch.Sph.", "Int. E", "Q&N", "Bhutan-A",
"Bhutan-C", "Bhutan-D", "Work-D", "Work-G", "EField", "Delta v")

# and plot
irt.Plot(ita.data.1$model, labs = labs)
irt.Plot(ita.data.2$model, labs = labs)
irt.Plot(ita.data.3$model, labs = labs)
}
