% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseAnalysisMCMR.R
\name{MCMR.response.analysis}
\alias{MCMR.response.analysis}
\title{MCMR response frequency and correlations}
\usage{
MCMR.response.analysis(data, booted = FALSE, nRuns = 5, makeCorPlot = FALSE)
}
\arguments{
\item{data}{Must be from the piql.data.select or data.select.Walkthrough functions}

\item{booted}{Logical (default = FALSE). FALSE means no bootstrapping will be
performed. TRUE turns on the bootstrapping feature.}

\item{nRuns}{Number of random samples to use in the bootstrapping (default = 5).
For publications it is recommended that 10,000 runs be performed since sample
error goes as 1/sqrt(nRuns).}

\item{makeCorPlot}{(Default = FALSE) If TRUE, then visualizations of the correlation
matrices will be produced. When booted = true, the mean correlations will be
plotted for each course.}
}
\value{
The response frequency and correlations for the MCMR
items on the assessment.
}
\description{
Calculates the response frequency and correlations for the MCMR
items on the assessment.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0)

# Get selection numbers and correlations
MCMR.response.analysis(temp.data)

# Can be booted
MCMR.response.analysis(temp.data, booted = TRUE)

# Can make plots of the correlations matrices
MCMR.response.analysis(temp.data, makeCorPlot = TRUE)
}
