% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleImputations.R
\name{multiple.imputations}
\alias{multiple.imputations}
\title{Apply multple imputation to given data sets}
\usage{
multiple.imputations(
  data,
  answerKey,
  remove.lessthan.50perc.answered = TRUE,
  plot.missing.pattern = FALSE,
  nImps = 5,
  nMaxIterations = 5
)
}
\arguments{
\item{data}{Can be either

1) An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions,

2) The output of piql.data.select with one or multiple courses selected.}

\item{answerKey}{Answer key for the assessment being analyzted. Generally can
be entered as something like pulled.data$answerKey from the pulldata function.}

\item{remove.lessthan.50perc.answered}{(Default = TRUE) Removed students who
answered less than 50 percent of the questions on the assessment. Removal
of these students is standard prectice in multiple imputation practice. If
you want to keep all students, set remove.lessthan.50perc.answered = TRUE.}

\item{plot.missing.pattern}{(Default = FALSE) If TRUE, then a plot of the
missing data patterns will be produced.}

\item{nImps}{(Dfault = 5) Number of imputed data sets to be created.}

\item{nMaxIterations}{(Default = 5) Number of iterations to be used to create
mean values used to estimate missing data. Larger numbers take longer to run,
but will give for consistent results.}
}
\value{
List of complete, imputed data sets for the given data formatted to be
readily used in all simplyPsycho functions. Suggestion, do not make plots of
this data using other functions. Get the results and then use the pool.imputations
function to get the pooled results for the imputed courses.

https://datascienceplus.com/imputing-missing-data-with-r-mice-package/
}
\description{
Uses multiple implutation to create nImps complete data sets.
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 2, MCMR.items = NA)
answerKey <- PIQLdata$answerkey

multiple.imputations(temp.data, answerKey)
}
