% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netACommunityMemberships.R
\name{netA.Community.Memberships}
\alias{netA.Community.Memberships}
\title{Community membership data for plotting}
\usage{
netA.Community.Memberships(
  data.num,
  cor.type = 1,
  lans.alpha = 0.05,
  nRuns = 100,
  printRunNumber = FALSE
)
}
\arguments{
\item{data.num}{An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions.}

\item{cor.type}{(default = 1). 1 = Pearson r, 2 = polychoric, 3 = partial correlation.}

\item{lans.alpha}{Cutoff value for the LANS sparcification procedure. Default is set to
0.05, since this acts similar to a p-value.}

\item{nRuns}{Number of random samples to use in the bootstrapping (default = 100).
For publications it is recommended that 10,000 runs be performed since sample
error goes as 1/sqrt(nRuns).}

\item{printRunNumber}{Default = FALSE. Print the run number currently being computed.
Useful when doing long runs and you want to keep track of there the code is.}
}
\value{
Community membership data to be put unto netA.Generate.PairedMat()
for the generation of heatmaps.
}
\description{
Generates community membership data using multiple network analysis
community detection methods and bootstrapping.
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0, MCMR.items = NA)
data.num <- temp.data$data.num

# Get community membership data
data.memberships <- netA.Community.Memberships(temp.data$data.num)
# Generate the pared matrix
data.PairedMat <- netA.Generate.PairedMat(thing.memberships$fastgreedy)
# Create heat map
heatmap(data.PairedMat)

}
