% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netAFilterLANS.R
\name{netA.Filter.LANS}
\alias{netA.Filter.LANS}
\title{Locally Adaptive Network Sparcification}
\usage{
netA.Filter.LANS(initial.network, lans.alpha)
}
\arguments{
\item{initial.network}{Initial network of the data. Output from graph.adjacency() function.}

\item{lans.alpha}{Cutoff value for the LANS sparcification procedure. Default is set to
0.05, since this acts similar to a p-value.}
}
\value{
A LANS sparcified network.
}
\description{
Takes in an undirected network and performs LANS on it.
}
