% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netAlansPlot.R
\name{netA.lans.plot}
\alias{netA.lans.plot}
\title{Locally Adaptive Network Sparcification Plot}
\usage{
netA.lans.plot(data.num, cor.type = 1, lans.alpha = 0.05)
}
\arguments{
\item{data.num}{An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions.}

\item{cor.type}{(default = 1). 1 = Pearson r, 2 = polychoric, 3 = partial correlation.}

\item{lans.alpha}{Cutoff value for the LANS sparcification procedure. Default is set to
0.05, since this acts similar to a p-value.}
}
\value{
A plot of the LANS sparcified network for the given data.
}
\description{
Takes in an undirected network and performs LANS on it.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPQLdata()
temp.piql.data <- PIQLdata$courses
data.num <- temp.data$data.num

# Get LANS plot of data
netA.lans.plot(temp.data$data.num, cor.type = 1)
}
