% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullData.R
\name{pulldata}
\alias{pulldata}
\title{Pull  data from AWS}
\usage{
pulldata(
  accessID,
  accessSecret,
  data.directory = "piqlgerqndata",
  name.assessment = NULL,
  region = "us-west-2"
)
}
\arguments{
\item{accessID}{AWS access ID.}

\item{accessSecret}{AWS secret access ID.}

\item{data.directory}{(Default = "piqlgerqndata") Name of the directory in AWS
the user is extracting data from.}

\item{name.assessment}{Name of the assessment who data is to be extracted. Default
option will result in a walk through helping the user select the correct data.}

\item{region}{(Default = "us-west-2") AWS region ID for database.}
}
\value{
A list of course data stored in AWS for the requested assessment.
}
\description{
Walks the user through pulling available data from AWS.
}
\examples{
# Pull in data from AWS and get some course data.
pulledData <- pulldata()
pulledData$courses # Course data
pulledData$answerkey # answer key
}
