% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piqlDataSelectandSplit.R
\docType{data}
\name{pulled.PIQL.data}
\alias{pulled.PIQL.data}
\title{Select specific course from the pulled PIQL data.}
\format{
An object of class \code{list} of length 4.
}
\usage{
pulled.PIQL.data
}
\arguments{
\item{pulled.PIQL.data}{Output from pullPQLdata}

\item{course}{(Default = 1 = 121) Course number (1 = 121, 2 = 122, 3 = 123, etc.) or position in course list:

Course: 121, 122, 123, 141, 142, 143, 224, 225, 226, 321, 322, 323, 325}

\item{MCMR.grading}{(Default = "Dichotomous") If undefined, then MCMR items will be graded dichotomously.

If "Selected", then selected = 1 and not selected = 0 for correct options on
the MCMR items will be performed.

If "FourScale", then MCMR items will be graded o the 4 scale:
(0 = completely incorrect, 1 = correct and incorrect, 2 = some correct, 3 = completely correct)}

\item{numBlanks.allowed}{(default = 0) Number of blanks allowed in the resulting data.}
}
\value{
A list of course data for the PIQL stored in AWS.

     $data.alpha ALPHABETICAL data for the selected course.

     $data.num NUMERICAL data for the selected course.

     $nS.details Details about how many students were removed due to the specified numBlanks.allowed value.
}
\description{
Selects a specific course from the pulled PIQL data and separates
the correct MCMR options marked 1 if they were selected and 0 if they were not.
}
\examples{
# Get PIQL data
PIQLdata <- pullPIQLdata()
temp.data <- piql.data.select(PIQLdata, MCMR.grading = "Selected", course = 1)
data.alpha <- temp.data$data.alpha
data.num <- temp.data$data.num
# Check number of student removed. Should be less than 10\%.
temp.data$nS.details

}
\keyword{datasets}
