% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullPIQLdata.R
\name{pullPIQLdata}
\alias{pullPIQLdata}
\title{Pull PIQL data from AWS}
\usage{
pullPIQLdata(accessID, accessSecret)
}
\value{
A list of course data for the PIQL stored in AWS.

     $courses course data as a list.

    $answerkey answer key stored in AWS.
}
\description{
Pulls PIQL data from AWS
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
PIQLdata$courses # Course data
PIQLdata$answerkey # answer key
}
