% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseFrequency.R
\name{response.Frequency}
\alias{response.Frequency}
\title{Response Frequency}
\usage{
response.Frequency(data, booted = FALSE, nRuns = 10, MCMR.separate = TRUE)
}
\arguments{
\item{data}{Can be either

1) An nS by nQ matrix or data frame of ALPHABETICAL data,
where nS is the number of students in the sample and nQ is the number of questions,

2) The output of piql.data.select with one or multiple courses selected.}

\item{booted}{Logical (default = FALSE). FALSE means no bootstrapping will be
performed. TRUE turns on the bootstrapping feature.}

\item{nRuns}{Number of random samples to use in the bootstrapping (default = 100).
For publications it is recommended that 10,000 runs be performed since sample
error goes as 1/sqrt(nRuns).}

\item{MCMR.separate}{(Default = TRUE). MCMR.separate = TRUE separates multiple selections
to get percentage for individual options on the MCMR items. MCMR.separate = FALSE
treats grouped selections as individual options.}
}
\value{
When booted = FALSE (the default setting) then the straight calculated
value will be returned. If booted = TRUE, the function will output the mean and
standard deviation for the CTT item difficulties calculated from nRuns randomly sampled
with replacement data sets from the given sample.
}
\description{
Calculates the frequency each response option is selected in the given data
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
data <- piql.data.select(PIQLdata, course = c(1,2), numBlanks.allowed = 0)

# Get Response frequencies for each item in each selected course
response.Frequency(data)
# Supports bootstrapping
response.Frequency(data, booted = TRUE)

}
