% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectQuestionFromData.R
\name{select.questions}
\alias{select.questions}
\title{Select specific course from the pulled PIQL data.}
\usage{
select.questions(selected.data, selected.questions = NULL)
}
\arguments{
\item{selected.data}{Output from a data selection function (like piql.data.select)}

\item{selected.questions}{A list of questions for each pulled course. Each course needs its own list
of questions. (NOT IMPLEMENTED YET)}
}
\value{
A list of course data filtered to include only the selected questions.

     $data.alpha ALPHABETICAL data for the selected course.

     $data.num NUMERICAL data for the selected course.

     $nS.details Details about how many students were removed due to the specified numBlanks.allowed value.
}
\description{
Selects a specific course from the pulled PIQL data and separates
the correct MCMR options marked 1 if they were selected and 0 if they were not.
}
\examples{
# Using Sample data in simplyPsycho
temp.data <- piql.data.select(simplySampleData, courses = c(1,2))
temp.data <- select.questions(temp.data)
}
