% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPsychoDatafile.R
\name{build.psycho.datafile}
\alias{build.psycho.datafile}
\title{Build simplyPsycho data from external data files. INTERNAL}
\usage{
build.psycho.datafile(
  datafile,
  new.data,
  new.course.name = NULL,
  new.term = NULL,
  answerKey = NULL,
  gen.blanks = FALSE,
  blank.ID = NULL
)
}
\description{
Build simplyPsycho data from external data files. INTERNAL
}
