% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttCronbachAlpha-Walkthrough.R
\name{cttCronbachAlpha.Walkthrough}
\alias{cttCronbachAlpha.Walkthrough}
\title{CTT Cronbach's Alpha WALKTHROUGH:}
\usage{
cttCronbachAlpha.Walkthrough(data)
}
\arguments{
\item{data}{Can be either

1) An nS by nQ matrix or data frame of a dichotomous graded (0 or 1) sample,
where nS is the number of students in the sample and nQ is the number of questions,

2) The output of piql.data.select with one or multiple courses selected.}
}
\value{
The Cronbach's Alpha calculation as specified in the walkthough.
}
\description{
Walks the user through the set up and use of calculating
the Cronbach's Alpha for an assessment using a given sample.
}
\examples{
# Use sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0, MCMR.items = NA)

# Begin walkthrough
cttCronbachAlpha.Walkthrough(temp.data)
}
