% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalCompareCourse.R
\name{gen.compare.courses}
\alias{gen.compare.courses}
\title{Compare courses}
\usage{
gen.compare.courses(
  data,
  makeBoxPlots = FALSE,
  makeDiffPlots = FALSE,
  makeSankeyPlots = FALSE,
  nBins = 7,
  nRuns = 100,
  return.gains = FALSE,
  makeGainsPlots = FALSE,
  ...
)
}
\arguments{
\item{data}{MUST BE the output of piql.data.select for multiple courses selected.}

\item{makeBoxPlots}{(Default = FALSE) If TRUE, then bar plots will be made for
all given courses in a single plot.

NOTE: Use the Left and Right arrow in the Plots window to switch between the plots.}

\item{makeDiffPlots}{(Default = FALSE) If TRUE, then point plots will be made for
comparing all possible course pairs.

NOTE: Use the Left and Right arrow in the Plots window to switch between the plots.}

\item{makeSankeyPlots}{(Default = FALSE) If TRUE, then Sankey plots will be made for
comparing all possible course pairs.

NOTE: Use the Left and Right arrow in the Viewer window to switch between the plots.}

\item{nBins}{Number of total score bins to use for the Sankey plots. Bins are roughly
"equal" ranges of total score. ~(number of total scores possible)/nBins}

\item{nRuns}{Number of runs to use when calculating error in CTT Difficulty.}

\item{return.gains}{(Default = FALSE) If FALSE, the gains are not returned for each student.
IF TRUE, then gains are reported for each student.}

\item{makeGainsPlots}{(Default = FALSE) If true, then plots of normalized gain/change versus
"pre" scores are created.}
}
\value{
Cohen's d with interpretation, p-value and adjusted p-value of difference in means,
course normalized gain, difference in item difficulty with error. On request
box plots, change in difficulty plots, Sankey plots, and gains versus pre score plots
cal all be created. Sankey plots and gains versus pre score plots require matched data.

NOTE: Use the Left and Right arrow in the Plots window to switch between the plots.
}
\description{
Use a t-test to compare total scores between two courses, and
ANOVA to compate three or more courses.

NOTE: Use the Left and Right arrow in the Plots window to switch between the plots.
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = c(1,2), numBlanks.allowed = 0, Matched = TRUE)
gen.compare.courses(temp.data,  makeSankeyPlots = TRUE)

gen.compare.courses(data)
## Box Plots can be made
gen.compare.courses(temp.data, makeBoxPlots = TRUE)
## Change in item difficulty pots can be made
gen.compare.courses(temp.data, makeDiffPlots = TRUE)

### MATCHED DATA
## Sankey plots can be made
gen.compare.courses(temp.data, makeSankeyPlots = TRUE)
## Plots of normalized gain/change versus pre score for each student
gen.compare.courses(temp.data, makeGainsPlots = TRUE)
## You can get the gain values by setting return.gains = TRUE
gen.compare.courses(temp.data, return.gains = TRUE)

}
