% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalSankeyPlot.R
\name{general.compare.courses.MakeSankey}
\alias{general.compare.courses.MakeSankey}
\title{Makes Sankey plot}
\usage{
general.compare.courses.MakeSankey(
  set1,
  set2,
  nQ,
  nBins = 7,
  couse.names,
  ii,
  jj
)
}
\arguments{
\item{data}{Can be either}
}
\description{
Makes Sankey plot for gen.compare.course
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
temp.data <- piql.data.select(PIQLdata, course = 2, numBlanks.allowed = 0)
data.num <- temp.data$data.num
}
