% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircPlot.R
\name{irc.plot}
\alias{irc.plot}
\title{Item Response Curve plotter:}
\usage{
irc.plot(irc.data, qq)
}
\arguments{
\item{irc.data}{Output from irc.get().}

\item{qq}{Question number you would like plotted.}
}
\value{
IRC plot for the requested item using the given irc.get() output.
}
\description{
Plots the result of irc.get().
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0, MCMR.items = NA)
data.alpha <- temp.data$data.alpha
data.num <- temp.data$data.num

# get irc data and plot item 2
irc.data <- irc.get(data.alpha, data.num)
irc.plot(irc.data, 2)
}
