% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleImputationsPool.R
\name{multiple.imputation.pooled}
\alias{multiple.imputation.pooled}
\title{Pooled Imputed results}
\usage{
multiple.imputation.pooled(to.be.pooled)
}
\arguments{
\item{to.be.pooled}{Must be the results of cttDifficulty, cttDiscrimination,
cttCronbachAlpha or cttpointBiserial.}
}
\value{
Mean and sd of the results for multiple imputed data sets.
}
\description{
Takes in results from simplyPsycho functions applied to multiply imputed
data sets and returns the mean and sd of the results.
}
\examples{
# Pull sample data
temp.data <- piql.data.select(simplySampleData, courses = 1, numBlanks.allowed = 0)
answerKey <- simplySampleData$answerkey

meh <- multiple.imputations(temp.data, answerKey)

thing <- cttDifficulty(meh)
#thing <- cttDiscrimination(meh)
#thing <- cttCronbachAlpha(meh)
#thing <- cttpointBiserial(meh)
}
