% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netAGeneratePairedMat.R
\name{netA.Generate.PairedMat}
\alias{netA.Generate.PairedMat}
\title{Create paired matrix for network analysis community detection heat maps}
\usage{
netA.Generate.PairedMat(memberMatrix)
}
\arguments{
\item{memberMatrix}{Output from netA.Community.Memberships() function.}
}
\value{
Paired matrix from community membership data to be plotted using heatmap().
}
\description{
Generates paired matrix from community membership data.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()
temp.data <- piql.data.select(PIQLdata, MCMR.grading = "Selected", course = 1)
data.num <- temp.data$data.num

# Get community membership data
data.memberships <- netA.Community.Memberships(temp.data$data.num)
# Generate the pared matrix
data.PairedMat <- netA.Generate.PairedMat(thing.memberships$fastgreedy)
# Create heat map
heatmap(data.PairedMat)

}
