% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piqlDataSelectandSplit.R
\name{piql.data.select}
\alias{piql.data.select}
\title{Select specific course from the pulled PIQL data.}
\usage{
piql.data.select(pulled.PIQL.data, courses = 1, numBlanks.allowed = 0)
}
\arguments{
\item{pulled.PIQL.data}{Output from pullPQLdata}

\item{numBlanks.allowed}{(default = 0) Number of blanks allowed in the resulting data.}

\item{course}{(Default = 1 = 121) Course number (1 = 121, 2 = 122, 3 = 123, etc.) or position in course list:

Course: 121, 122, 123, 141, 142, 143, 224, 225, 226, 321, 322, 323, 325

*Can be a list of courses.*}
}
\value{
A list of course data for the PIQL stored in AWS.

     $data.alpha ALPHABETICAL data for the selected course.

     $data.num NUMERICAL data for the selected course.

     $nS.details Details about how many students were removed due to the specified numBlanks.allowed value.
}
\description{
Selects a specific course from the pulled PIQL data.
}
\examples{
# Pull in PIQL data from AWS and get some course data.
PIQLdata <- pullPIQLdata()

# Get answerkey
answerKey <- PIQLdata$answerkey

# Pull course data and split into alpha and numerical variables
temp.data <- piql.data.select(PIQLdata, course = 1)
data.alpha <- temp.data$data.alpha
data.num <- temp.data$data.num

# Check number of student removed. Should be less than 10\%.
temp.data$nS.details
}
